refactor: change folder of the date component belonging to the card
