+++ src/frontend/src/components/Card/Common/Date/styles.module.scss
@@ -1,4 +1,4 @@
-@use "../../../styles/variables" as v;
+@use "../../../../styles/variables" as v;
 
 .container {
   &__title {
